rename: model 하위 패키지 이름 수정
- 복합단어로 이루어진 패키지 이름 수정
