+++ src/main/java/com/workout/model/postcomment/PostCommentUpdateParam.java
@@ -1,4 +1,4 @@
-package com.workout.model.post_comment;
+package com.workout.model.postcomment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
